smarter URL detection
