+++ packages/uploadthing/test/get-full-url.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { getFullApiUrl } from "../src/client";
+import { getFullApiUrl } from "../src/internal/get-full-api-url";
 
 describe("getFullApiUrl", () => {
   it("should return the provided url if it is already absolute", () => {
